Fix the OOM that killed the first slot run
Exit status 137, oom-kill in a dataloader worker, 48 GB exhausted in 15
minutes without completing an epoch.

The dataset handed back up to 64 masks per image as float32 at full
resolution -- 50 MB per image, ~1.6 GB per batch of 32 -- and 16 workers
each prefetching 4 batches put ~64 of those in flight at once. The click
model never came near this because its target is a single mask, 0.75 MB.

Three changes, largest first.

Training masks are now built at the loss's own resolution rather than the
image's. The loss immediately downsampled full-resolution targets to the
logits' grid anyway, so carrying them at full size was not just expensive
but pointless. Validation keeps full resolution deliberately: its IoU is
the number reported against the click model's 0.6194, and measuring it on
a coarser grid would quietly flatter it.

Masks are stored as bool rather than float32, a quarter of the bytes,
with a single cast in the loss where the copy is cheap.

Workers drop 16 -> 8 and prefetch 4 -> 2, and pinning is off. The click
model's settings were tuned when an item was one small mask; here workers
dominate memory rather than throughput.

Together: 50 MB per training image down to 2.5 MB, a 20x reduction.

build_loader now prints its worst-case mask memory at startup, so the
next time this budget is wrong it is visible in the first seconds of the
log rather than inferred from an exit code 15 minutes later.

Also cached upsampled masks per slot in SlotPrediction. Validation asks
about every object in an image against one encoding, and the same slots
rank repeatedly, so this turns an interpolation per instance into one per
slot.

+++ src/training/losses.py
@@ -198,13 +198,16 @@ class HungarianMaskLoss(nn.Module):
             if target.numel() == 0 or target.shape[0] == 0:
                 continue
 
+            # Cast once, here: the dataset stores masks as bool to keep them
+            # out of the dataloader's memory budget, and interpolate needs float.
+            target = target.float()
             if target.shape[-2:] != (height, width):
                 target = F.interpolate(
                     target.unsqueeze(1), size=(height, width), mode="nearest"
                 ).squeeze(1)
 
             flat_logits = mask_logits[i].flatten(1)  # (N, P)
-            flat_target = target.flatten(1).float()  # (K, P)
+            flat_target = target.flatten(1)          # (K, P)
 
             with torch.no_grad():
                 cost = self._cost_matrix(flat_logits.detach(), flat_target)
